# debug-swagger-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Financiera Backend API",
    "version": "1.0.0",
    "description": "API for vehicle sales platform. Provides endpoints for user authentication, vehicle references management, user administration, credit application management, and fingerprint/attendance processing."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Versión 1 de la API"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "integer",
            "description": "HTTP status code"
          }
        },
        "required": [
          "error",
          "code"
        ]
      },
...
      "Usuario": {
        "type": "object",
        "properties": {
          "idUsuario": {
            "type": "integer",
            "description": "Unique user ID"
          },
          "nombre": {
            "type": "string",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "telefono": {
            "type": "string",
            "description": "User phone number"
          },
          "direccion": {
            "type": "string",
            "description": "User address"
          },
          "idRol": {
            "type": "integer",
            "description": "Role ID (1 = Admin, 2 = Vendedor, 3 = Cliente)"
          },
          "estado": {
            "type": "string",
            "enum": [
              "activo",
              "inactivo"
            ],
            "description": "User status"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "summary": "Listar todos los usuarios",
        "description": "Devuelve una lista de todos los usuarios registrados. Solo accesible para administradores.",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                },
                "example": [
                  {
                    "idUsuario": 1,
                    "nombre": "Juan Pérez",
                    "email": "juan@financiera.com",
                    "telefono": 1234567890,
                    "direccion": "Calle Falsa 123",
                    "idRol": 1,
                    "estado": "activo"
                  }
                ]
              }
            }
          },
          "403": {
            "description": "Acceso denegado (no administrador)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "No tienes permisos para realizar esta acción",
                  "code": 403
                }
              }
            }
          },
          "429": {
            "description": "Demasiadas solicitudes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Demasiadas solicitudes, intenta de nuevo en 15 minutos",
                  "code": 429
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo usuario (solo administradores)",
        "description": "Crea un nuevo usuario con un rol específico.\nRoles disponibles: 2 = Vendedor, 3 = Cliente.\nSolo accesible para administradores.\n",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre completo del usuario"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Correo electrónico único"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "Contraseña (mínimo 8 caracteres)"
                  },
                  "idRol": {
                    "type": "integer",
                    "description": "Rol del usuario (2 = Vendedor, 3 = Cliente)",
                    "enum": [
                      2,
                      3
                    ]
                  }
                },
                "required": [
                  "nombre",
                  "email",
                  "password",
                  "idRol"
                ]
              },
              "example": {
                "nombre": "María Gómez",
                "email": "maria@financiera.com",
                "password": "SecurePass123!",
                "idRol": 3
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                },
                "example": {
                  "idUsuario": 2,
                  "nombre": "María Gómez",
                  "email": "maria@financiera.com",
                  "idRol": 3,
                  "estado": "activo"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado (no administrador)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "No tienes permisos para realizar esta acción",
                  "code": 403
                }
              }
            }
          },
          "409": {
            "description": "Correo ya registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "El correo ya está registrado",
                  "code": 409
                }
              }
            }
          },
          "429": {
            "description": "Demasiadas solicitudes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Demasiadas solicitudes, intenta de nuevo en 15 minutos",
                  "code": 429
                }
              }
            }
          }
        }
      }
    },
...
    "/api/usuarios/register": {
      "post": {
        "summary": "Registrar un nuevo usuario",
        "description": "Registra un nuevo cliente en la plataforma.\nNo requiere autenticación previa.\nEl usuario registrado tendrá el rol de cliente (idRol = 3) por defecto.\nLímite de 5 registros por hora.\n",
        "tags": [
          "Usuarios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre completo del usuario"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Correo electrónico único"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "Contraseña (mínimo 8 caracteres, con letras y números)"
                  }
                },
                "required": [
                  "nombre",
                  "email",
                  "password"
                ]
              },
              "example": {
                "nombre": "Ana López",
                "email": "ana.lopez@financiera.com",
                "password": "SecurePass123!"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                },
                "example": {
                  "idUsuario": 2,
                  "nombre": "Ana López",
                  "email": "ana.lopez@financiera.com",
                  "idRol": 3,
                  "estado": "activo"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "El correo electrónico no es válido",
                  "code": 400
                }
              }
            }
          },
          "409": {
            "description": "Correo ya registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "El correo ya está registrado",
                  "code": 409
                }
              }
            }
          },
          "429": {
            "description": "Demasiados intentos de registro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Demasiados intentos, intenta de nuevo en 15 minutos",
                  "code": 429
                }
              }
            }
          }
        }
      }
    },
...
  }
}